Fixed documentation

+++ required-libs/GPIO.cpp
@@ -20,8 +20,8 @@ void (*pint_irq_ptr[PINT_INPUT_COUNT])(void);
  *
  * @retval None.
  */
-Pin::Pin(uint32_t pin_index, const uint32_t mode) {
-
+Pin::Pin(uint8_t pin_index, const uint8_t mode) {
+	/* Call Pin initialization */
 	init(pin_index, mode);
 }
 
@@ -36,9 +36,10 @@ Pin::Pin(uint32_t pin_index, const uint32_t mode) {
  *
  * @retval None.
  */
-Pin::Pin(uint32_t pin_index, const uint32_t mode, const uint32_t initial_value) {
-
+Pin::Pin(uint8_t pin_index, const uint8_t mode, const uint8_t initial_value) {
+	/* Call Pin initialization */
 	init(pin_index, mode);
+	/* Set initial value */
 	GPIO->B[port][pin] = initial_value;
 }
 
@@ -52,30 +53,21 @@ Pin::Pin(uint32_t pin_index, const uint32_t mode, const uint32_t initial_value) 
  *
  * @retval None.
  */
-void Pin::init(uint32_t pin_index, const uint32_t mode) {
-
+void Pin::init(uint8_t pin_index, const uint8_t mode) {
+	/* Get the pin index */
 	pin = pin_index;
+	/* Get the PORT number */
 	port = getPORT(pin);
-
-	// Enable GPIO CLK
-	if (port == 1) {
-		SYSCON->SYSAHBCLKCTRL0 |= SYSCON_SYSAHBCLKCTRL0_GPIO1_MASK;
-	}
-	else {
-		SYSCON->SYSAHBCLKCTRL0 |= SYSCON_SYSAHBCLKCTRL0_GPIO0_MASK;
-	}
-
-	// Get the actual pin index within the port
-	if (pin > 31) {
-		pin -= 32;
-	}
-
-	if (mode == INPUT) {
-		GPIO->DIR[port] &= ~(1UL << pin);
-	}
-	else {
-		GPIO->DIR[port] |= 1UL << pin;
-	}
+	/* Enable clock for GPIO1 */
+	if (port == 1) { SYSCON->SYSAHBCLKCTRL0 |= SYSCON_SYSAHBCLKCTRL0_GPIO1_MASK; }
+	/* Enable clock for GPIO0 */
+	else { SYSCON->SYSAHBCLKCTRL0 |= SYSCON_SYSAHBCLKCTRL0_GPIO0_MASK; }
+	/* Adjust, if necessary, the pin index value */
+	if (pin > 31) { pin -= 32; }
+	/* Initialize Pin as INPUT */
+	if (mode == INPUT) { GPIO->DIR[port] &= ~(1UL << pin); }
+	/* Initialize Pin as OUTPUT */
+	else { GPIO->DIR[port] |= 1UL << pin; }
 }
 
 /*!
@@ -89,17 +81,18 @@ void Pin::init(uint32_t pin_index, const uint32_t mode) {
  *
  * @retval None.
  */
-void Pin::attachInterrupt(void (*f)(void), pint_pin_enable_t pint_mode, uint32_t pint_pin_index) {
-
-	uint32_t pin_index = (port == 0)? pin : pin + 32;		// Get the original pin index
-
-	SYSCON_AttachSignal(SYSCON, (pint_pin_int_t)pint_pin_index, getSYSCON_CONNECTION(pin_index));		// Connect the pin to the
-																										// proper PINT channel
-
-	pint_irq_ptr[pint_pin_index] = f;	// Save the function pointer
-
+void Pin::attachInterrupt(void (*f)(void), pint_pin_enable_t pint_mode, uint8_t pint_pin_index) {
+	/* Get the original pin index */
+	uint8_t pin_index = (port == 0)? pin : pin + 32;
+	/* Connect the Pin to the proper PINT channel */
+	SYSCON_AttachSignal(SYSCON, (pint_pin_int_t)pint_pin_index, getSYSCON_CONNECTION(pin_index));
+	/* Store the function pointer */
+	pint_irq_ptr[pint_pin_index] = f;
+	/* Enable PINT module */
 	PINT_Init(PINT);
+	/* PINT configuration */
 	PINT_PinInterruptConfig(PINT, (pint_pin_int_t)pint_pin_index, pint_mode, pint_callback);
+	/* Enable callback */
 	PINT_EnableCallbackByIndex(PINT, (pint_pin_int_t)pint_pin_index);									// Enable trigger by channel
 }
 
@@ -115,9 +108,10 @@ void Pin::attachInterrupt(void (*f)(void), pint_pin_enable_t pint_mode, uint32_t
  *
  * @retval None.
  */
-void pint_callback(pint_pin_int_t pintr, uint32_t pmatch_status) {
-
+void pint_callback(pint_pin_int_t pintr, uint8_t pmatch_status) {
+	/* Check if there is a function for that index */
 	if(pint_irq_ptr[pintr]) {
+		/* Call function */
 		pint_irq_ptr[pintr]();
 	}
 }
